Move `Weekday` to domain
Moves the `Weekday` enum from the `flow_engine` module to the `domain` module to improve code organization and reusability.

This change centralizes the `Weekday` definition, making it accessible to other parts of the application that may need to work with weekdays.
