+++ src/flow_engine/expression.rs
@@ -1,5 +1,5 @@
-use crate::domain::Number;
 use crate::domain::property::{BooleanProperty, NumberProperty, PropertyType};
+use crate::domain::{Number, Weekday};
 use crate::flow_engine::Context;
 use crate::flow_engine::expression::ExpressionError::UnknownProperty;
 use Weekday::*;
@@ -80,35 +80,6 @@ impl WeekdayCondition {
             WeekdayCondition::Weekdays => vec![Monday, Tuesday, Wednesday, Thursday, Friday],
             WeekdayCondition::Weekend => vec![Saturday, Sunday],
         }
-    }
-}
-
-#[derive(PartialEq, Eq, Hash, Debug, Clone)]
-pub enum Weekday {
-    Monday,
-    Tuesday,
-    Wednesday,
-    Thursday,
-    Friday,
-    Saturday,
-    Sunday,
-}
-
-impl Weekday {
-    pub fn as_index(&self) -> usize {
-        match self {
-            Monday => 0,
-            Tuesday => 1,
-            Wednesday => 2,
-            Thursday => 3,
-            Friday => 4,
-            Saturday => 5,
-            Sunday => 6,
-        }
-    }
-
-    pub fn all() -> [Weekday; 7] {
-        [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
     }
 }
